Make the project a gradle repo
Although I didn't want to make this a gradle project as most of the things I'll use are going to be from the scala library, it got messier to even run the unit tests without a dependency management tool.
So, to keep the development process simple, decided to add gradle.

We only need two dependencies for now. scala language and scalatest for unit testing.